pyproject.toml
[tool.poetry]
name = "parcerias-ufpa"
version = "0.1.0"
description = ""
authors = ["Thiago Leonardo <lthiago2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
sqlmodel = "^0.0.37"
psycopg2-binary = "^2.9.11"
pydantic-settings = "^2.13.1"
email-validator = "^2.3.0"
asyncpg = "^0.31.0"
bcrypt = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.22"
pandas = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
